Class BaseModel created
Some methods moved from BasePage to BaseModel

Tests from MainPageTest moved to DashboardTest

MainPageTest class deleted
